change the count we send to the buffer again
i'm really not sure what i'm looking at. thee color and the verts
are way off..but i can see a resembelance of the original image...so
it's a good start. I'm thinking it's go to do with my offsets to the
buffer info i'm sending...or to the vert/tex calcs..