awesome, chatbox is getting there. hack by using a tab widget for scrolling
:)

i didn't think i'd be able to do it
somewhat easily. i was encountering difficulties with making a text
history for the chat with a scrollbar..it kept wanting to scroll the
button and textbox as well..even the example does so, which i don't see
where that would even be desireable.

but this solution seems to be what i need..