for now, we will ignore the dedicated server scenario
(it's pretty trivial to add, simply make Game only create the server
based on arguments to binary), and so the Client technically creates
the server if he chooses to host and play (SP or host MP), but of course
only when the user chooses to either play SP or host and play an MP
server will it do so. So we won't make assumptions based on that (validity
of m_server indicates this).

In the event that the client and the server will exist, the client will
handle the event loop of the server too.

athe tricky part is going to be handling collision and stuff through
the World...and also splitting out the rendering, so the spritesheetmanager
is not called if in server-only mode. luckily that was already modularized
so that part shouldn't be too bad. mostly the collision i need to think
about, and corresponding update code