hm, for a lot of these methods i should be passing around a Player
which actually incidentally contains a Camera. Players now aren't just
dumb representations of what the client wants to do, but in a way..
are the clients themselves. From the server's perspective when not
communicating, that is. This enables culling visible entities before being
packeted out. Of course, it means that the server needs to know the camera
position of each and every client..which on second thought isn't as bad
of an idea as it may sound, I believe. At some stage the server *has*
to know what the player is looking at anyways..so what better way to do it?