set default sizes for objects created, default sizes are set by the renderer
when we register entities with them..i wouldn't think we'd need to otherwise.
there may me scenarios where we want to change size (increase, mostly),
after initialization, and this should handle that reasonable well.

the default sprite size is based on the frame that it loads. maybe it won't
scale, but it's suitable for now, and was only 6 SLOC

so now spriteshheet description (YAML) parsing works now, so it knows
which texcoords to use for each frame name, instead of just using the whole
spritesheet