clear the initial spawn area for the player properly now.
calculate the area better (still a fucking hack), but a better one.

also, turns out my entity position and entity physics body positions need
a lot of improvement/unification, as presently..it turns out, they
are two separate things that evidently need to be set. e.g. when i make
a player object i need to tell it to make the phys body at that position
as well as set the player position itself.