hopefully remove some hardcoded resolution stuff and use settings instead
this shows that i have *a lot* of work to do when it comes to handling
multiple resolutions.. sigh, and i'm not sure how to properly handle it.

terraria handles it in an interesting way..as far as most are concerned
terraria's res is static...but it expands the range of the viewport..
this royally fucks over PvP apparently. i've no idea how to handle this
nicely..